requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
google-auth==2.49.2
gunicorn==25.3.0
idna==3.11
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
python-http-client==3.3.7
requests==2.33.1
sendgrid==6.12.5
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
whitenoise==6.12.0
